cloud_functions/gmail_sync/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.13.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0

